third_party/WebKit/LayoutTests/inspector/components/color-expected.txt

Tests Common.Color

Dumping 'red' in different formats:
 - rgb(255, 0, 0)
 - rgba(255, 0, 0, 1)
 - hsl(0, 100%, 50%)
 - hsla(0, 100%, 50%, 1)
 - hsv(0, 100%, 100%)
 - hsva(0, 100%, 100%, 1)
 - #ff0000
 - #f00
 - red
 - default: red
 - inverse color: rgba(0, 255, 255, 1)
 - setAlpha(0.42): rgba(255, 0, 0, 0.42)
Dumping 'green' in different formats:
 - rgb(0, 128, 0)
 - rgba(0, 128, 0, 1)
 - hsl(120, 100%, 25%)
 - hsla(120, 100%, 25%, 1)
 - hsv(120, 100%, 50%)
 - hsva(120, 100%, 50%, 1)
 - #008000
 - null
 - green
 - default: green
 - inverse color: rgba(255, 127, 255, 1)
 - setAlpha(0.42): rgba(0, 128, 0, 0.42)
Dumping 'blue' in different formats:
 - rgb(0, 0, 255)
 - rgba(0, 0, 255, 1)
 - hsl(240, 100%, 50%)
 - hsla(240, 100%, 50%, 1)
 - hsv(240, 100%, 100%)
 - hsva(240, 100%, 100%, 1)
 - #0000ff
 - #00f
 - blue
 - default: blue
 - inverse color: rgba(255, 255, 0, 1)
 - setAlpha(0.42): rgba(0, 0, 255, 0.42)
Dumping 'cyan' in different formats:
 - rgb(0, 255, 255)
 - rgba(0, 255, 255, 1)
 - hsl(180, 100%, 50%)
 - hsla(180, 100%, 50%, 1)
 - hsv(180, 100%, 100%)
 - hsva(180, 100%, 100%, 1)
 - #00ffff
 - #0ff
 - cyan
 - default: cyan
 - inverse color: rgba(255, 0, 0, 1)
 - setAlpha(0.42): rgba(0, 255, 255, 0.42)
Dumping 'magenta' in different formats:
 - rgb(255, 0, 255)
 - rgba(255, 0, 255, 1)
 - hsl(300, 100%, 50%)
 - hsla(300, 100%, 50%, 1)
 - hsv(300, 100%, 100%)
 - hsva(300, 100%, 100%, 1)
 - #ff00ff
 - #f0f
 - magenta
 - default: magenta
 - inverse color: rgba(0, 255, 0, 1)
 - setAlpha(0.42): rgba(255, 0, 255, 0.42)
Dumping 'yellow' in different formats:
 - rgb(255, 255, 0)
 - rgba(255, 255, 0, 1)
 - hsl(60, 100%, 50%)
 - hsla(60, 100%, 50%, 1)
 - hsv(60, 100%, 100%)
 - hsva(60, 100%, 100%, 1)
 - #ffff00
 - #ff0
 - yellow
 - default: yellow
 - inverse color: rgba(0, 0, 255, 1)
 - setAlpha(0.42): rgba(255, 255, 0, 0.42)
Dumping 'white' in different formats:
 - rgb(255, 255, 255)
 - rgba(255, 255, 255, 1)
 - hsl(0, 0%, 100%)
 - hsla(0, 0%, 100%, 1)
 - hsv(0, 0%, 100%)
 - hsva(0, 0%, 100%, 1)
 - #ffffff
 - #fff
 - white
 - default: white
 - inverse color: rgba(0, 0, 0, 1)
 - setAlpha(0.42): rgba(255, 255, 255, 0.42)
Dumping 'black' in different formats:
 - rgb(0, 0, 0)
 - rgba(0, 0, 0, 1)
 - hsl(0, 0%, 0%)
 - hsla(0, 0%, 0%, 1)
 - hsv(0, 0%, 0%)
 - hsva(0, 0%, 0%, 1)
 - #000000
 - #000
 - black
 - default: black
 - inverse color: rgba(255, 255, 255, 1)
 - setAlpha(0.42): rgba(0, 0, 0, 0.42)
Dumping 'rgb(94, 126, 91)' in different formats:
 - rgb(94, 126, 91)
 - rgba(94, 126, 91, 1)
 - hsl(115, 16%, 43%)
 - hsla(115, 16%, 43%, 1)
 - hsv(115, 28%, 49%)
 - hsva(115, 28%, 49%, 1)
 - #5e7e5b
 - null
 - null
 - default: rgb(94, 126, 91)
 - inverse color: rgba(161, 129, 164, 1)
 - setAlpha(0.42): rgba(94, 126, 91, 0.42)
Dumping 'rgba(94, 126, 91, 0.5)' in different formats:
 - null
 - rgba(94, 126, 91, 0.5)
 - null
 - hsla(115, 16%, 43%, 0.5)
 - hsv(115, 28%, 49%)
 - hsva(115, 28%, 49%, 0.5)
 - null
 - null
 - null
 - default: rgba(94, 126, 91, 0.5)
 - inverse color: rgba(161, 129, 164, 0.5)
 - setAlpha(0.42): rgba(94, 126, 91, 0.42)
Dumping 'hsl(212, 55%, 32%)' in different formats:
 - rgb(37, 79, 126)
 - rgba(37, 79, 126, 1)
 - hsl(212, 55%, 32%)
 - hsla(212, 55%, 32%, 1)
 - hsv(212, 71%, 50%)
 - hsva(212, 71%, 50%, 1)
 - #254f7e
 - null
 - null
 - default: hsl(212, 55%, 32%)
 - inverse color: rgba(218, 176, 129, 1)
 - setAlpha(0.42): rgba(37, 79, 126, 0.42)
Dumping 'hsla(212, 55%, 32%, 0.5)' in different formats:
 - null
 - rgba(37, 79, 126, 0.5)
 - null
 - hsla(212, 55%, 32%, 0.5)
 - hsv(212, 71%, 50%)
 - hsva(212, 71%, 50%, 0.5)
 - null
 - null
 - null
 - default: hsla(212, 55%, 32%, 0.5)
 - inverse color: rgba(218, 176, 129, 0.5)
 - setAlpha(0.42): rgba(37, 79, 126, 0.42)
Dumping '#00FFFF' in different formats:
 - rgb(0, 255, 255)
 - rgba(0, 255, 255, 1)
 - hsl(180, 100%, 50%)
 - hsla(180, 100%, 50%, 1)
 - hsv(180, 100%, 100%)
 - hsva(180, 100%, 100%, 1)
 - #00FFFF
 - #0ff
 - cyan
 - default: #00ffff
 - inverse color: rgba(255, 0, 0, 1)
 - setAlpha(0.42): rgba(0, 255, 255, 0.42)
Dumping '#0FF' in different formats:
 - rgb(0, 255, 255)
 - rgba(0, 255, 255, 1)
 - hsl(180, 100%, 50%)
 - hsla(180, 100%, 50%, 1)
 - hsv(180, 100%, 100%)
 - hsva(180, 100%, 100%, 1)
 - #00ffff
 - #0FF
 - cyan
 - default: #0ff
 - inverse color: rgba(255, 0, 0, 1)
 - setAlpha(0.42): rgba(0, 255, 255, 0.42)
